requirements.txt
# Voice Chat Bridge Dependencies

# STT Engines
vosk>=0.3.45
funasr-onnx>=0.4.0

# TTS Engines
edge-tts>=6.1.0
pyttsx3>=2.90

# Configuration
pydantic>=2.6.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Audio Processing
librosa>=0.10.0
scipy>=1.11.0

# Optional (for future extensions)
requests>=2.31.0
# aiohttp>=3.9.0
